backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
supabase==2.3.4
sqlalchemy==2.0.25

# Data processing
pandas==2.2.0
numpy==1.26.3

# HTTP requests
httpx==0.26.0
requests==2.31.0

# Environment variables
python-dotenv==1.0.1

# Scheduling
apscheduler==3.10.4

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML (optional - add as needed)
openai==1.10.0
langchain==0.1.5

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
